Adds proxysql helm chart
This chart is based on dysnix/charts/tree/main/dysnix/proxysql with many
additions made by KlickTipp in our fork.

We finally made the decision to take the helm chart to our repository, as our team is maintaining the chart for a long time now, and it's easier to have in our own repo.